% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sh601398}
\alias{sh601398}
\title{工商银行(601398)数据}
\format{数据格式为 data frame，有8个变量:
\describe{
  \item{date}{交易日期}
  \item{open}{开盘价}
  \item{high}{最高价}
  \item{low}{最低价}
  \item{close}{收盘价}
  \item{volume}{交易量}
  \item{amount}{交易金额}   
  \item{factor}{复权因子}
  ...
}}
\source{
\url{http://finance.sina.com.cn/stock/}
}
\usage{
sh601398
}
\description{
数据来源自新浪股票，包含复权因子。
}
\keyword{datasets}

